# samples/web-csr/dressca-backend/api-docs/web-consumer/api-specification.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "ECサイトDressca",
    "title": "Dressca",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/catalog-items": {
      "get": {
        "description": "カタログアイテムを検索して返します。",
        "operationId": "getByQuery",
        "parameters": [
          {
            "in": "query",
            "name": "brandId",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedListOfCatalogItemResponse"
                }
              }
            },
            "description": "成功。"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "リクエストエラー。"
          }
        },
        "summary": "カタログアイテムを検索して返します。",
        "tags": [
          "CatalogItems"
        ]
      }
    },
...
    "/api/orders": {
      "post": {
        "description": "買い物かごに登録されている商品を注文します。",
        "operationId": "postOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "成功。"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "リクエストエラー。"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "サーバーエラー。"
          }
        },
        "summary": "買い物かごに登録されている商品を注文します。",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{orderId}": {
      "get": {
        "description": "注文情報を取得します。",
        "operationId": "getById",
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "成功。"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "注文 ID が存在しない。"
          }
        },
        "summary": "注文情報を取得します。",
        "tags": [
          "Orders"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AccountResponse": {
        "type": "object",
        "properties": {
          "consumptionTax": {
            "type": "number"
          },
          "consumptionTaxRate": {
            "type": "number",
            "format": "double"
          },
          "deliveryCharge": {
            "type": "number"
          },
          "totalItemsPrice": {
            "type": "number"
          },
          "totalPrice": {
            "type": "number"
          }
        },
        "required": [
          "consumptionTax",
          "consumptionTaxRate",
          "deliveryCharge",
          "totalItemsPrice",
          "totalPrice"
        ]
      },
...
      "CatalogItemResponse": {
        "type": "object",
        "properties": {
          "assetCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "catalogBrandId": {
            "type": "integer",
            "format": "int64"
          },
          "catalogCategoryId": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "productCode": {
            "type": "string"
          }
        },
        "required": [
          "catalogBrandId",
          "catalogCategoryId",
          "description",
          "id",
          "name",
          "price",
          "productCode"
        ]
      },
      "CatalogItemSummaryResponse": {
        "type": "object",
        "properties": {
          "assetCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "productCode": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "productCode"
        ]
      },
...
      "OrderItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "itemOrdered": {
            "$ref": "#/components/schemas/CatalogItemSummaryResponse"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "subTotal": {
            "type": "number"
          },
          "unitPrice": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "quantity",
          "subTotal",
          "unitPrice"
        ]
      },
      "OrderResponse": {
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/AccountResponse"
          },
          "azanaAndOthers": {
            "type": "string"
          },
          "buyerId": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderDate": {
            "type": "string",
            "format": "date-time"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponse"
            }
          },
          "postalCode": {
            "type": "string"
          },
          "shikuchoson": {
            "type": "string"
          },
          "todofuken": {
            "type": "string"
          }
        },
        "required": [
          "azanaAndOthers",
          "buyerId",
          "fullName",
          "id",
          "orderDate",
          "postalCode",
          "shikuchoson",
          "todofuken"
        ]
      },
      "PagedListOfCatalogItemResponse": {
        "type": "object",
        "properties": {
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogItemResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PostOrderRequest": {
        "type": "object",
        "properties": {
          "azanaAndOthers": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "shikuchoson": {
            "type": "string"
          },
          "todofuken": {
            "type": "string"
          }
        },
        "required": [
          "azanaAndOthers",
          "fullName",
          "postalCode",
          "shikuchoson",
          "todofuken"
        ]
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
    }
  }
}